feat: 프론트에게 줄 위도 경도 데이터 double로 반환

+++ src/main/java/com/selfbell/criminal/dto/CriminalCoordDto.java
@@ -10,7 +10,7 @@ import java.math.BigDecimal;
 @Builder
 public class CriminalCoordDto {
     private final String address;   // 예: "서울특별시 광진구 중곡동 18-109"
-    private final BigDecimal lat;   // 위도 (latitude)
-    private final BigDecimal lon;   // 경도 (longitude)
-    private final long distanceMeters; // 사용자와의 거리(m). nearby에서만 채워짐
+    private final double lat;   // 위도 (latitude)
+    private final double lon;   // 경도 (longitude)
+    private final double distanceMeters; // 사용자와의 거리(m). nearby에서만 채워짐
 }
